Linting + GUI fixed
`flakehell` and `black` are used for formatting consistently.
`pyproject.toml` has the configuration used. `requirements-dev.txt`
has been updated with the requirements.

GUI was updated and generally works. There are still bugs
that need to be worked out. The GUI and CLI of the program
run the same underlining functions and generally will work the same.