GUI Funationality Added
Added a module for the GUI. The layout has been changed. There is
a device type option to choose which device you are working with.
It defaults to the first device in the list. The artifacts in the list
will be reloaded automatically.

It has not be completely tested yet.